check ffmpeg before rendering video
