+++ scripts/process_data.py
@@ -16,7 +16,7 @@ import tyro
 from rich.console import Console
 from rich.progress import track
 
-from nerfstudio.utils import colmap_utils
+from nerfstudio.utils import colmap_utils, install_checks
 
 CONSOLE = Console(width=120)
 
@@ -32,44 +32,6 @@ CAMERA_MODELS = {
     "perspective": CameraModel.OPENCV,
     "fisheye": CameraModel.OPENCV_FISHEYE,
 }
-
-
-def check_ffmpeg_installed():
-    """Checks if ffmpeg is installed."""
-    ffmpeg_path = shutil.which("ffmpeg")
-    if ffmpeg_path is None:
-        CONSOLE.print("[bold red]Could not find ffmpeg. Please install ffmpeg.")
-        print("See https://ffmpeg.org/download.html for installation instructions.")
-        print("ffmpeg is only necissary if using videos as input.")
-        sys.exit(1)
-
-
-def check_colmap_installed():
-    """Checks if colmap is installed."""
-    colmap_path = shutil.which("colmap")
-    if colmap_path is None:
-        CONSOLE.print("[bold red]Could not find COLMAP. Please install COLMAP.")
-        print("See https://colmap.github.io/install.html for installation instructions.")
-        sys.exit(1)
-
-
-def get_colmap_version(default_version=3.8) -> float:
-    """Returns the version of COLMAP.
-    This code assumes that colmap returns a version string of the form
-    "COLMAP 3.8 ..." which may not be true for all versions of COLMAP.
-
-    Args:
-        default_version: Default version to return if COLMAP version can't be determined.
-    Returns:
-        The version of COLMAP.
-    """
-    output = run_command("colmap", verbose=False)
-    assert output is not None
-    for line in output.split("\n"):
-        if line.startswith("COLMAP"):
-            return float(line.split(" ")[1])
-    CONSOLE.print(f"[bold red]Could not find COLMAP version. Using default {default_version}")
-    return default_version
 
 
 def get_vocab_tree() -> Path:
@@ -234,7 +196,7 @@ def run_colmap(
         verbose: If True, logs the output of the command.
     """
 
-    colmap_version = get_colmap_version()
+    colmap_version = install_checks.get_colmap_version()
 
     (colmap_dir / "database.db").unlink(missing_ok=True)
 
@@ -408,8 +370,8 @@ def main(
         verbose: If True, print extra logging.
     """
 
-    check_ffmpeg_installed()
-    check_colmap_installed()
+    install_checks.check_ffmpeg_installed()
+    install_checks.check_colmap_installed()
 
     output_dir.mkdir(parents=True, exist_ok=True)
     image_dir = output_dir / "images"
