+++ scripts/render.py
@@ -28,6 +28,7 @@ from nerfstudio.cameras.camera_paths import get_path_from_json, get_spiral_path
 from nerfstudio.cameras.cameras import Cameras
 from nerfstudio.configs.base_config import Config  # pylint: disable=unused-import
 from nerfstudio.pipelines.base_pipeline import Pipeline
+from nerfstudio.utils import install_checks
 from nerfstudio.utils.eval_utils import eval_setup
 from nerfstudio.utils.rich_utils import ItersPerSecColumn
 
@@ -115,6 +116,8 @@ class RenderTrajectory:
             double_nerfacto_nerf_samples=self.double_nerfacto_nerf_samples,
             eval_num_rays_per_chunk=self.eval_num_rays_per_chunk,
         )
+
+        install_checks.check_ffmpeg_installed()
 
         seconds = self.seconds
